Some refactoring to move check_jobs and documentation.
